requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
click==8.2.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
frozenlist==1.7.0
future==1.0.0
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
yt-dlp==2025.6.9
